\name{famSKATRC}
\alias{famSKATRC}

\docType{package}
\title{
\packageTitle{famSKATRC}
}
\description{
\packageDescription{famSKATRC}
}

\author{
\packageAuthor{famSKATRC}

Maintainer: \packageMaintainer{famSKATRC}
}
\references{
Saad M and Wijsman EM (2014) Combining family- and population-based imputation data for association analysis of rare and common variants in large pedigrees.
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4190076/
}
\keyword{famSKATRC}
\keyword{package}
\keyword{imputation}
\keyword{rare}
\keyword{common}
\keyword{GWAS}
